Split fill and morph code into separate modules
Put common constants and helper functions into a
separate module. Makes it easier to get an overview
of the individual disconnected sections of the fill
process.

Disable multiprocess morph ops for win32

factor out multiprocess setup and add
try-catch w. fallback to single-process
morphing on failure.

Improve documentation
